Moving Miro transforming logic
