+++ catalogue_pipeline/transformer/src/main/scala/uk/ac/wellcome/transformer/transformers/CalmTransformableTransformer.scala
@@ -5,5 +5,5 @@ import uk.ac.wellcome.models.{CalmTransformable, Work}
 import scala.util.Try
 
 class CalmTransformableTransformer extends TransformableTransformer[CalmTransformable] {
-  override def transform(transformable: CalmTransformable): Try[Work] = ???
+  override def transform(transformable: CalmTransformable): Try[Option[Work]] = ???
 }
